Rename aioconsole.code to aioconsole.console
